Jsystem work (#399)
* Almost match JASBasicInst

* JASDrumSet OK

* Some JAudio2 struct adjustment

* Match JKRHeap::callAllDisposer

* JAISe OK

* Fix setAutoDelete

* JAISound OK

* Match a few functions in linklist

* Few matches in d_a_e_ym

* JUTTexture OK

* Work on JUTGamePad

* Import JUTResource

* Work on JUTResFont::getFontCode

+++ libs/JSystem/JUtility/JUTCacheFont.cpp
@@ -479,7 +479,7 @@ asm ResFONT* JUTResFont::getResFont() const {
 #pragma push
 #pragma optimization_level 0
 #pragma optimizewithasm off
-asm u16 JUTResFont::getFontType() const {
+asm int JUTResFont::getFontType() const {
     nofralloc
 #include "asm/JSystem/JUtility/JUTCacheFont/getFontType__10JUTResFontCFv.s"
 }
